Start rewriting functions
There was a problem with pasting FBX files using NVil's native clipboard
feature in Wine+Mono environment on a GNU/Linux operating system.
Proably something with Mono not fully supporting routines that are
responsible for that. I contacted the developer of NVil and he was kind
enough to implement two new tools that allow to programatically export
and import files. The goal of this rewrite is to accomodate this
program to take advantage of those functions instead of the clipboard.

The second goal is to separate instructions for GNU/Linux and Windows as
they will now be somewhat different than when this AppLink used NVil's
clipboard feature. I'll probably still use clipboard feature for Windows
import/export functions.